refactor(types): swap legacy-path imports to rllm.types
Apply the same import swap to the files that were held back in the
previous pass — concrete agents under rllm/agents/, AgentExecutionEngine,
two environments modules, and rllm/trainer/deprecated/. They are slated
to move to rllm/v0/ in a follow-up PR, but updating imports now means
the v0 move becomes a pure relocation rather than a relocation + import
fix-up.

After this, every remaining `from rllm.agents.agent import …` line in
the package imports only BaseAgent — the ABC stays put until the v0
move.
