+++ rllm/agents/math_agent.py
@@ -1,7 +1,8 @@
 import copy
 from typing import Any
 
-from rllm.agents.agent import Action, BaseAgent, Step, Trajectory
+from rllm.agents.agent import BaseAgent
+from rllm.types import Action, Step, Trajectory
 
 
 class MathAgent(BaseAgent):
