+++ rllm/trainer/deprecated/tinker_agent_trainer.py
@@ -21,7 +21,6 @@ import torch
 from omegaconf import OmegaConf
 from transformers import AutoTokenizer
 
-from rllm.agents.agent import Episode, Step, Trajectory
 from rllm.engine.agent_execution_engine import AsyncAgentExecutionEngine
 from rllm.trainer.deprecated.tinker_metrics_utils import (
     compute_training_metrics,
@@ -29,6 +28,7 @@ from rllm.trainer.deprecated.tinker_metrics_utils import (
     print_metrics_table,
 )
 from rllm.trainer.deprecated.tinker_policy_trainer import TinkerPolicyTrainer
+from rllm.types import Episode, Step, Trajectory
 
 if TYPE_CHECKING:
     pass
