+++ rllm/trainer/deprecated/tinker_workflow_trainer.py
@@ -16,11 +16,11 @@ import tinker
 import torch
 from transformers import AutoTokenizer
 
-from rllm.agents.agent import Episode
 from rllm.engine.agent_workflow_engine import AgentWorkflowEngine
 from rllm.engine.rollout.tinker_engine import TinkerEngine
 from rllm.trainer.deprecated.tinker_agent_trainer import TinkerAgentTrainer
 from rllm.trainer.deprecated.tinker_policy_trainer import TinkerPolicyTrainer
+from rllm.types import Episode
 
 if TYPE_CHECKING:
     pass
